+++ packages/welcome/lib/changelog-view.js
@@ -50,39 +50,16 @@ export default class ChangeLogView {
             <p>Feel free to read our <a href="https://github.com/pulsar-edit/pulsar/blob/master/CHANGELOG.md">Full Change Log</a>.</p>
             <ul>
               <li>
-                Pulsar v1.127.1 Hotfix: Reverted a Wayland-related change that Linux users reported issues with on Electron 12. (The remaining changes listed below are from Pulsar v1.127.0.)
+                Vastly improved support for right-to-left text editing — meaning languages like Arabic, Hebrew, and Farsi.
               </li>
               <li>
-                Added a Jasmine 2-based test runner, migrated core editor tests to use it. Packages bundled into the core editor can migrate their tests to use this as well, over time. The Jasmine 1 test runner remains available.
+                Added support for comments in JSON — both via a new grammar for <code>.jsonc</code> files and via a setting to enable comments in regular <code>.json</code> files.
               </li>
               <li>
-                Added <code>--enable-features=UseOzonePlatform</code> and <code>--ozone-platform=wayland</code> as parameters when running under Wayland on Linux (avoids using xwayland, which causes rendering problems on some systems, especially with NVidia)
+                Fixed performance issues that can arise when reopening a project with existing editor windows.
               </li>
               <li>
-                Many Tree-sitter/parser/grammar improvements.
-                <ul>
-                  <li>
-                    Updated to <code>web-tree-sitter</code> version <code>0.25.3</code>.
-                  </li>
-                  <li>
-                    Fixed a bug preventing folds from updating after code changes in some scenarios.
-                  </li>
-                  <li>
-                    Better folding behavior in Python.
-                  </li>
-                  <li>
-                    Better folding and syntax highlighting in Ruby of <code>case</code>/<code>in</code> statements.
-                  </li>
-                  <li>
-                    Better syntax highlighting of private members in JavScript.
-                  </li>
-                  <li>
-                    Better folding of multiline comments in PHP.
-                  </li>
-                </ul>
-              </li>
-              <li>
-                Updated the `read` dependency in ppm
+                Updated DOMPurify. This is a security fix.
               </li>
             </ul>